% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CowbellAnalyzer.R
\docType{data}
\name{allFun}
\alias{allFun}
\title{allFun: Data set for Fluency, Absorption, Fit and Fun.}
\description{
Data set of two games containing measurements of Fluency, Absorption, Fit (= Perceived difficulty) and Fun.
Fluency and Absorption are measured on a scale from 1..7, Fun and Fit are measured  1..9.
}
