% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CowbellAnalyzer.R
\name{print.cowbellConcept}
\alias{print.cowbellConcept}
\title{Summarizes the cowbell concept with the formula and value ranges.}
\usage{
\method{print}{cowbellConcept}(x, ...)
}
\arguments{
\item{x}{Object generated with function \code{\link{generateCowbellConcept}}}

\item{...}{Unused for compatibility only.}
}
\description{
Implementation of the \code{\link{print}} generic.
}
\examples{
# Generate a concept and display it.
concept<-generateCowbellConcept(Fun ~ Fluency * Absorption, 1, 9, 1, 7, 1, 7)
concept
}
\seealso{
\code{\link{generateCowbellConcept}}
}
